feat: Cria policy de Brand

+++ app/Http/Controllers/BrandController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\BrandUpdateRequest;
 use App\Models\Brand;
 use App\Services\BrandServices;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class BrandController extends Controller
 {
@@ -19,6 +20,8 @@ class BrandController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Brand::class);
+
         $brands = $this->brandServices->list();
 
         return response()->json($brands);
@@ -29,6 +32,8 @@ class BrandController extends Controller
      */
     public function store(BrandStoreRequest $request)
     {
+        Gate::authorize('create', Brand::class);
+
         $brand = $this->brandServices->store($request);
 
         return response()->json(['message' => 'Brand created successfully', 'brand' => $brand]);
@@ -39,6 +44,8 @@ class BrandController extends Controller
      */
     public function show(Brand $brand)
     {
+        Gate::authorize('view', $brand);
+
         return response()->json($brand);
     }
 
@@ -47,6 +54,8 @@ class BrandController extends Controller
      */
     public function update(BrandUpdateRequest $request, Brand $brand)
     {
+        Gate::authorize('update', $brand);
+
         $brand = $this->brandServices->update($request, $brand);
 
         return response()->json(['message' => 'Brand updated successfully', 'brand' => $brand]);
@@ -57,6 +66,8 @@ class BrandController extends Controller
      */
     public function destroy(Brand $brand)
     {
+        Gate::authorize('delete', $brand);
+
         $brand->delete();
 
         return response()->json(['message' => 'Brand deleted successfully']);
